% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accr-package.R
\docType{package}
\name{accr-package}
\alias{accr}
\alias{accr-package}
\title{accr: Analyzing Accelerometer Data for Small Animal Studies}
\description{
A toolkit for analyzing accelerometer data from small animal studies. The package provides a range of functions and methods to process, visualize, and extract meaningful information from raw acceleration signals and ODBA (Overall Dynamic Body Acceleration) collected from animal-borne accelerometers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpio-be/accr}
  \item Report bugs at \url{https://github.com/mpio-be/accr/issues}
}

}
\author{
\strong{Maintainer}: Mihai Valcu \email{mvalcu@gwdg.de} (\href{https://orcid.org/0000-0002-6907-7802}{ORCID})

Authors:
\itemize{
  \item Anne Aulsebrook \email{anne.aulsebrook@bi.mpg.de} (\href{https://orcid.org/0000-0002-9217-7043}{ORCID})
}

}
\keyword{internal}
