% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actogram.R
\name{actogram.ODBA}
\alias{actogram.ODBA}
\title{actogram
actogram}
\usage{
\method{actogram}{ODBA}(O, threshold, ...)
}
\arguments{
\item{threshold}{a number specifying a threshold or a function see [accr::resting_threshold()]}

\item{...}{}
}
\description{
actogram
actogram
}
\examples{
data(pesaODBA)
setODBA(pesaODBA, "datetime_", "ODBA")
actogram(pesaODBA, threshold = resting_threshold, sigma = 2 )
Fixed threshold based on 100 individuals from the same cohort as the pesaODBA male.
actogram(pesaODBA, threshold = 780.83)


}
