% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_cpp}
\alias{dist_cpp}
\title{Function to calculate euclidean distance between two vectors}
\usage{
dist_cpp(ptr1, ptr2, index1, index2, length)
}
\arguments{
\item{ptr1}{Rcpp::XPtr of recording object 1}

\item{ptr2}{Rcpp::Xptr of recording object 2}

\item{index1}{start index for ptr1 (R notation)}

\item{index2}{start index for ptr2 (R notation)}

\item{length}{total length to calculate}
}
\value{
A numeric vector of length one
}
\description{
Function to calculate euclidean distance between two vectors
}

