% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_r}
\alias{dist_r}
\title{Calculate Euclidean distance between two large matrices}
\usage{
dist_r(m1, m2, n1 = 1, n2 = 1, len = max(0, min(nrow(m1) - n1, nrow(m2) -
  n2)), chn = min(ncol(m1), ncol(m2)), chunksize = 1e+05)
}
\arguments{
\item{m1,m2}{matrix objects or objects of matrix shape}

\item{n1,n2}{Starting index for m1 and m2}

\item{len}{len to be calculated for both m1 and m2}

\item{chunksize}{divide computation to chunks}
}
\value{
a numeric value, euclidean distance
}
\description{
Calculate Euclidean distance between two large matrices
}
\examples{
d <- dist_r(r1, r2, 1, 1, 1000)
}
\author{
Siyu Sun
}

