% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resting.R
\name{resting_threshold}
\alias{resting_threshold}
\title{resting_threshold}
\usage{
resting_threshold(x, method = "mixfit", ...)
}
\arguments{
\item{x}{An ODBA data.table. See [accr::setODBA()]}

\item{method}{Only "mixfit" available at the moment. See [mixR::mixfit]}

\item{...}{Further arguments passed to the underlining methods.}
}
\value{
A value for the resting threshold is given.
}
\description{
resting_threshold
}
\examples{
data(pesaODBA)
setODBA(pesaODBA, "datetime_", "ODBA")
thr = resting_threshold(pesaODBA, method = "mixfit")
density(pesaODBA$ODBA) |> plot()
abline(v = thr, col = 2)


}
