% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{setODBA}
\alias{setODBA}
\title{setODBA
setODBA}
\usage{
setODBA(d, datetime_, ODBA, threshold = NA)
}
\arguments{
\item{d}{data.table}

\item{datetime_}{datetime}

\item{ODBA}{'character'}

\item{threshold}{function or value}
}
\description{
setODBA
setODBA
}
