% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dist.R
\name{dist_r}
\alias{dist_r}
\title{Calculate Euclidean distance between two signals}
\usage{
dist_r(raw1, raw2, index1 = 1, index2 = 1, length = min(nrow(raw1),
  nrow(raw2)) - 1, chunksize = 1e+05)
}
\arguments{
\item{raw1,raw2}{Rawdata objects}

\item{index1,index2}{Starting index for raw1 and raw2}

\item{length}{length to be calculated for both raw1 and raw2}

\item{chunksize}{divide computation to chunks}
}
\value{
a numeric value, euclidean distance
}
\description{
Calculate Euclidean distance between two signals
}
\examples{
d <- dist_r(r1, r2, 1, 1, 1000)
}
\author{
Siyu Sun
}

