% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resting.R
\name{resting_threshold}
\alias{resting_threshold}
\title{resting_threshold}
\usage{
resting_threshold(x, method = "mixfit", Sigma = 2)
}
\arguments{
\item{x}{An ODBA data.table. See [accr::setODBA()]}

\item{method}{Only "mixfit" available at the moment. See [mixR::mixfit]}

\item{Sigma}{SD units away away from the mean to set the resting cutoff. Default t0 2.}
}
\value{
A value for the resting threshold is given.
}
\description{
resting_threshold
}
\examples{
data(pesaODBA)
setODBA(pesaODBA, "datetime_", "ODBA")
thr = resting_threshold(pesaODBA)
density(pesaODBA$ODBA) |> plot()
abline(v = thr, col = 2)

}
